Adds - Build-logic module for reusing build configurations between modules - Hilt in app module - Jetpack compose in app module - Kotlin test for unit tests - Jacoco for test coverage - Github workflows to run unit tests
makes sure app builds
