+++ .github/ci-gradle.properties
@@ -1,0 +1,28 @@
+#
+# Copyright 2020 The Android Open Source Project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+org.gradle.daemon=false
+org.gradle.parallel=true
+org.gradle.workers.max=2
+
+kotlin.incremental=false
+kotlin.compiler.execution.strategy=in-process
+
+# Controls KotlinOptions.allWarningsAsErrors.
+# This value used in CI and is currently set to false.
+# If you want to treat warnings as errors locally, set this property to true
+# in your ~/.gradle/gradle.properties file.
+warningsAsErrors=false
