+++ app/src/main/java/com/thermondo/androidchallenge/ThermondoApp.kt
@@ -1,0 +1,7 @@
+package com.thermondo.androidchallenge
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class ThermondoApp : Application()
