+++ build-logic/convention/src/main/kotlin/AndroidApplicationConventionPlugin.kt
@@ -1,8 +1,6 @@
 import com.android.build.api.dsl.ApplicationExtension
-import com.rodrigoguerrero.buildlogic.convention.configureAndroidApplication
-import com.rodrigoguerrero.buildlogic.convention.configureAndroidLint
-import com.rodrigoguerrero.buildlogic.convention.configureDetekt
-import com.rodrigoguerrero.buildlogic.convention.configureKotlinAndroid
+import com.thermondo.buildlogic.convention.configureAndroidApplication
+import com.thermondo.buildlogic.convention.configureKotlinAndroid
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
@@ -13,14 +11,11 @@ class AndroidApplicationConventionPlugin : Plugin<Project> {
             with (pluginManager) {
                 apply("com.android.application")
                 apply("org.jetbrains.kotlin.android")
-                apply("io.gitlab.arturbosch.detekt")
             }
 
             extensions.configure<ApplicationExtension> {
                 configureAndroidApplication(this)
                 configureKotlinAndroid(this)
-                configureAndroidLint(this)
-                configureDetekt(this)
             }
         }
     }
