+++ build-logic/convention/src/main/kotlin/AndroidLibraryComposeConventionPlugin.kt
@@ -1,6 +1,5 @@
 import com.android.build.gradle.LibraryExtension
-import com.rodrigoguerrero.buildlogic.convention.configureAndroidCompose
-import com.rodrigoguerrero.buildlogic.convention.configureDetekt
+import com.thermondo.buildlogic.convention.configureAndroidCompose
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
@@ -11,12 +10,10 @@ class AndroidLibraryComposeConventionPlugin : Plugin<Project> {
         with (target) {
             with(pluginManager) {
                 apply("com.android.library")
-                apply("io.gitlab.arturbosch.detekt")
             }
 
             extensions.configure<LibraryExtension> {
                 configureAndroidCompose(this)
-                configureDetekt(this)
             }
         }
     }
