+++ build-logic/convention/src/main/kotlin/AndroidLibraryConventionPlugin.kt
@@ -1,8 +1,6 @@
 import com.android.build.gradle.LibraryExtension
-import com.rodrigoguerrero.buildlogic.convention.AppConfig
-import com.rodrigoguerrero.buildlogic.convention.configureAndroidLint
-import com.rodrigoguerrero.buildlogic.convention.configureDetekt
-import com.rodrigoguerrero.buildlogic.convention.configureKotlinAndroid
+import com.thermondo.buildlogic.convention.AppConfig
+import com.thermondo.buildlogic.convention.configureKotlinAndroid
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalogsExtension
@@ -17,13 +15,10 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
             with (pluginManager) {
                 apply("com.android.library")
                 apply("org.jetbrains.kotlin.android")
-                apply("io.gitlab.arturbosch.detekt")
             }
 
             extensions.configure<LibraryExtension> {
                 configureKotlinAndroid(this)
-                configureAndroidLint(this)
-                configureDetekt(this)
                 defaultConfig.targetSdk = AppConfig.targetSdk
             }
