+++ build-logic/convention/src/main/kotlin/AndroidRoomLibraryConventionPlugin.kt
@@ -1,6 +1,5 @@
 import com.android.build.gradle.LibraryExtension
-import com.rodrigoguerrero.buildlogic.convention.configureAndroidRoom
-import com.rodrigoguerrero.buildlogic.convention.configureDetekt
+import com.thermondo.buildlogic.convention.configureAndroidRoom
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
@@ -11,12 +10,10 @@ class AndroidRoomLibraryConventionPlugin : Plugin<Project> {
         with (target) {
             with (pluginManager) {
                 apply("org.jetbrains.kotlin.kapt")
-                apply("io.gitlab.arturbosch.detekt")
             }
 
             extensions.configure<LibraryExtension> {
                 configureAndroidRoom(this)
-                configureDetekt(this)
             }
         }
     }
