--- build-logic/convention/src/main/kotlin/com/rodrigoguerrero/buildlogic/convention/KotlinAndroid.kt
@@ -1,48 +1,0 @@
-package com.rodrigoguerrero.buildlogic.convention
-
-import com.android.build.api.dsl.CommonExtension
-import org.gradle.api.JavaVersion
-import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPluginExtension
-import org.gradle.kotlin.dsl.configure
-import org.gradle.kotlin.dsl.withType
-import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-internal fun Project.configureKotlinAndroid(
-    commonExtension: CommonExtension<*, *, *, *, *>,
-) {
-    commonExtension.apply {
-        compileSdk = AppConfig.compileSdk
-
-        defaultConfig {
-            minSdk = AppConfig.minSdk
-        }
-        compileOptions {
-            sourceCompatibility = AppConfig.javaVersion
-            targetCompatibility = AppConfig.javaVersion
-        }
-        configureKotlin()
-    }
-}
-
-internal fun Project.configureKotlinJvm() {
-    extensions.configure<JavaPluginExtension> {
-        sourceCompatibility = AppConfig.javaVersion
-        targetCompatibility = AppConfig.javaVersion
-    }
-
-    configureKotlin()
-}
-
-
-private fun Project.configureKotlin() {
-    tasks.withType<KotlinCompile>().configureEach {
-        kotlinOptions {
-            freeCompilerArgs = freeCompilerArgs + listOf(
-                "-opt-in=kotlin.RequiresOptIn",
-            )
-            jvmTarget = JavaVersion.VERSION_17.toString()
-            allWarningsAsErrors = true
-        }
-    }
-}
