+++ build-logic/convention/src/main/kotlin/com/thermondo/buildlogic/convention/AndroidApplication.kt
@@ -1,0 +1,21 @@
+package com.thermondo.buildlogic.convention
+
+import com.android.build.api.dsl.ApplicationExtension
+
+internal fun configureAndroidApplication(
+    applicationExtension: ApplicationExtension
+) {
+    applicationExtension.apply {
+        defaultConfig {
+            applicationId = AppConfig.applicationId
+            targetSdk = AppConfig.targetSdk
+            versionName = AppConfig.versionName
+            versionCode = AppConfig.versionCode
+            testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+
+            vectorDrawables {
+                useSupportLibrary = true
+            }
+        }
+    }
+}
