+++ build.gradle.kts
@@ -1,7 +1,19 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
-@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
-plugins {
-    alias(libs.plugins.com.android.application) apply false
-    alias(libs.plugins.org.jetbrains.kotlin.android) apply false
+buildscript {
+    repositories {
+        google()
+        mavenCentral()
+    }
+
+    dependencies {
+        classpath(libs.com.android.tools.build.gradle)
+        classpath(libs.org.jetbrains.kotlin.gradle.plugin)
+        classpath(libs.com.google.dagger.hilt.android.gradle.plugin)
+        classpath(libs.app.cash.paparazzi.gradle.plugin)
+        classpath(libs.org.jacoco.core)
+    }
 }
-true // Needed to make the Suppress annotation work for the plugins block
+
+tasks.register("clean", Delete::class) {
+    delete(rootProject.buildDir)
+}
